requirements.txt
exceptiongroup==1.2.1
iniconfig==2.0.0
packaging==24.1
pdf2image==1.17.0
pillow==10.3.0
pluggy==1.5.0
pytesseract==0.3.10
pytest==8.2.2
tomli==2.0.1
